+++ catnip/_commit.py
@@ -1,1 +1,1 @@
-__commit__ = "85ba19e"
+__commit__ = "7810ab3"